SyncEngine.sync now returns a Promise
