+++ unnamedcast/LoginViewController.swift
@@ -23,11 +23,9 @@ class LoginViewController: UIViewController {
     APIClient().request(ep).then { _, _, user in
       NSUserDefaults.standardUserDefaults().setObject(user.id, forKey: "user_id")
     }.then { () -> Void in
-      let ds = SyncEngine()
-      ds.sync {
-        dispatch_async(dispatch_get_main_queue()) {
-          self.performSegueWithIdentifier("Login2Main", sender: self)
-        }
+      let engine = SyncEngine()
+      engine.sync().then {
+        self.performSegueWithIdentifier("Login2Main", sender: self)
       }
     }
   }
